Move csv and functional to utils package
This package is not exactly cohesive... The idea is to hold all the core
functionality that can be used on both Node and Scriptable.

Use rollup-plugin-ts instead of rollup-plugin-esbuild so we can produce
.d.ts (etc.) files.

Also, some minimal testing! Originally I had planned to use ts-jest, but
it was quite slow since it used tsc by default. I didn't want add babel
as a new dependency (although Jest already uses it a bit) and have
ts-jest use that.

So, switch to a esbuild for Jest. There are packages for this, but most
of them seems a bit off putting. A comment in a GitHub issue for one of
them pointed out that you can roll your own esbuild-using Jest
transformer with a quite small amount of code. The jest-customizer is
simplified enough now that it is almost not worth having it as a runtime
component, but I'll keep it in place for now in case ts-jest adopts
esbuild and we can switch back to it.