isbn: if no results and only a single fault, reject with it
The developer will always have to check for both a rejection and the
fault lists. Not sure which should be prioritized.

And, except for the first few isbn/work checks, the fetched data
determines whether there can be a single or multiple faults (multiple
work keys may be rare, but seems possible based on the schema, but
multiple editions will be common -- the whole reason for this
functionality).