GetISBNs: also pull ISBN (-10) from data
In my test data all the records with ISBN10 also have ISBN13, but not
all the records with ISBN13 have ISBN10. This would make sense if there
were some non-978 ISBN-13s, but they are all 978-prefixed. So, they all
have equivalents, but Goodreads did not supply some of the ISBN-10s. The
next step will address that.