GetISBNs: derive both ISBNs
The test data I have been using mostly has both for all the entires that
have either, but a few are missing ISBN-10 values (despite having
978-prefixed ISBN-13 values). For whichever ISBN variant we pull out of
the data, derive the equivalent -10 and -13 variants.

Since the this will definitely create duplicates (if any row has both
ISBN and ISBN13 columns), send the final data through a Set to remove
duplication. In addition to the primary intra-item, source/derived
duplication, this will also squash duplication across items if somehow
two different items had the same (or equivalent) ISBNs. Not much point
in having such duplicates (except as a flag that there is a potential
flaw in the data that GR supplied?), so no great loss.