utils: do not bundle, node: only bundle
It probably only makes sense to bundle the dependencies at the "app"
level, so stop doing that in `utils`, which is a library.

Also, stop producing non-bundled node "app" files.

The CommonJS version doesn't work because everything is imported with
`require()` and we are using some ESM-only modules now (p-limit and
p-throttle). There must be a way to have rollup generate dynamic imports
for these ESM-only modules, but I haven't found it yet.

The ESM version almost works, but it would require changing

    import { parse, unparse } from 'papaparse';

to

    import pp from 'papaparse';
    const { parse, unparse } = pp;

to satisfy Node, whose CommonJS exports detection heuristics do not
recognize the method that papaparse uses to populate its module.exports.

Changing the syntax isn't really a big deal, but if the unbundled CJS
version is broken anyway, might as drop the unbundled ESM version, too.