cacheEditionsPromisor: new 'editions of' cache
This cache is fancier in that it normalizes its argument and resolution
ISBNs, and modifies the resolution set so that it is reflexive,
symmetric, and transitive (X is always an edition of X; when X is an
edition of Y, then Y is an edition of X; when X is an edition of Y and Y
is an edition of Z, then X is an edition of Z).

If the data sources are well maintained, this should already be the
case, but we go ahead and arrange it ourselves to try to limit how many
queries we will make against external services. For inconsistent data
sources, this might end up limiting the data that we gather. This is
something that random-ish expiration and cross-service comparison might
help with.