cachePromisor: caller does validation
Stop requiring that validators be passed in when restoring a save cache.
Either hand back the same thing we gave you, or validate it yourself.

Provide marshalling and unmarshalling hooks when saving and restoring.
But really, since the SavedCache type is not opaque, all that mapping
could be outsourced completely...