otherEditionsOfISBN: reclassify faults
The old classification was mostly useless. The caller probably didn't
care which phase of the process the fault came from. Instead classify
the faults as warnings or temporary faults. Warnings are maybe useful to
know about, but generally did not cause major problems. Temporary faults
are for unexpected errors that might be fixed eventually.

This kind of classification (probably along with a new "permanent"
classification) should be helpful when the cache is able to do
expiration, and we would like to expire results with temporary faults
sooner than those with just warnings.

Temporary means "try again later, but not right away" (i.e. cache for a
"short" time).

 * data not `JSON.parse`-able
 * top level structural validation error
 * empty array holding critical data (work.works, editions.entries)
 * no workIDs found at all
 * no ISBNs found at all
 * a rejection or thrown error occurred

Warnings can be ignored, lettings the cache entry live for a "normal"
period of time. There may be warnings and temporary faults at the same
time. The temporary presence of temporary faults is more important than
the presence of warnings.

 * validation error on work.works[n]
 * validation error on editions.entries[n]
 * no isbns from an editions.entries[n]