cache: push CacheControl union into fn's return type
It seemed "cleaner" if we specified that the supplied fn should return
strictly `Promise<T>` or `Promise<CacheControl<T>>`, but that causes
type inferencing problems if fn is built by a function that returns a
`.then`-mapped value (e.g. to convert an EditionsISBNResult to
Set<string>).

    const oe = otherEditionsOfISBN(fetcher);
    const cached = cachePromisor(isbn=>oe(isbn).then(r=>r.isbns));

So, switch to accepting a function returning
`Promise<T|CacheControl<T>>` instead. It is less precise in that the
function could randomly return either a `T` or a `CacheControl<T>` but
it keeps the type checker happy in the above case.

It also is a better reflection of what the code actually accepts (we can
and do check the type of the resolved value at runtime, but we can't
actually check the function's return type at runtime (before calling it
and and letting its returned Promise resolve)).