rollup: watch options
- do not clear the screen before a build
- wait a short time (20 milliseconds) before building

Occasionally (especially for the utils package) a build will happen
immediately after an expected build. Probably VS Code is writing a file
twice (or just not atomicly?).

Not clearing the screen makes it more obvious that is has happened, and
the delay seems to help prevent it. I binary-searched between 0 and 1000
and picked a "round" value just a bit higher than where the problem
seems to recur.