editions: OpenLibrary_Search
I found Python isbnlib (https://github.com/xlcnd/isbnlib) using this.

https://openlibrary.org/dev/docs/api/search

It is not quite as documented (uhh, `editions.json` isn't totally
documented either), but it seems to get what I want in a single request.
One of the example ISBNs in the Open Library API docs has so many
editions that `editions.json` is paginated several times. But this
search with `fields=isbn` seems to return all of the ISBNs without any
pagination (although the `start` and `offset` properties seem to hint
that it might happen in some circumstances).