rebundle papaparse to make it Node-ESM importable
When using (e.g.) `import {parse} from 'papaparse'`, TypeScript handled
it okay (because `@types/papaparse` tells it what is exported?), but
Node would refuse to import it because it the CJS-only `papaparse` did
not conform to Node's heuristic-based export detection.

Use Rollup (and its CommonJS plugin) to rebundle `papaparse` in its own
package. This lets us be sure the exports are Node-ESM-parsable (we
should be using the generated ESM file, but even if we end up using the
CJS portion, the rewritten exports should be understood by Node's ESM
import mechanism).