editions: Fetcher can CacheControl error results
Also, pass cache duration back through EditionsISBNResults so that
getISBNs can use it as a cache duration instead of the default "had
fault" duration.

This is useful for 429 and 503 HTTP status with a Retry-After header.

It probably makes sense to let Fetcher also CacheControl the main string
response, but that is a more invasive change and none of the services
seem to give any cache-related headers (e.g. Cache-Control or Expires).