asEditionsISBNResults: remove withISBNs parameter
The only place where the set wasn't actually ISBNs was when gathering
the work IDs in the Open Library WorkEditions. But there, and all the
other places where we used falsy `withISBNs`, the set was already
checked to be empty, so it was never a problem to include them.

This functionality was a holdover from when `EditionsISBNResults.isbn`
itself optional.