jest: jest-watch-suspend for better control
In watch mode, Jest was running much more often than I would like. For
example, when updating something in utils, it would run when the file
was saved, then again when rollup finished building in utils.

Sometimes Jest seems to just get stuck in a loop running tests, and
sometimes it crashes while a Git rebase is in progress. These may or may
not be helped as much by suspending.