rollup: top level config
A small configuration change in the individual package's typescript
plugin allows the various rollup package configurations to be gathered
together at the top level (with a bit of not-completely-general path
rewriting). This works in watch mode and should be better than the "one
watcher per changing package" that I have been using.