Scriptable tool: more progress: active, waiting, fetched
The most useful is fetched. This lets the user see that something is
still happening even if a bunch of OpenLibrary WorkEditions requests
have filled the queue.

OL:WE queries are a UI problem because they require at least two
fetches. The extra fetches are scheduled (by the throttle) to run after
the ones that have already "queued". So, if we are only using OL:WE, we
will see 10 requests start right away, but we won't see any finish until
the 11th fetch has finished (i.e. wee see a repeating pattern: 10
seconds of no query progress, 10 seconds of finishing queries, ...).

Showing fetch finishes at least gives the user a 1-per-second update
that something is still happening.