Scriptable tool: use separate cache file
This moves towards supporting an alternate cache store that is used
(along with the fake fetcher) only for testing (since the fake fetcher
provides obviously bogus data, it should be kept separate from real
data).

Also, the cache can be cleared without affecting the saved UI data.
Which is not that big of a deal, since there isn't much saved UI data
(it isn't too arduous to reselect the command, shelf and GetISBNs
options).