tool-core: change beforeAndAfter implementation
Something about the old version of `beforeAndAfter` seems to cause
problems with Jest when supplying a reporter combined with a fetcher
that throws (or rejects). Not sure why the new code behaves differently.

The problem seemed to stem from unhandled rejections. Maybe it was
happening because Jest only asks Node for `unhandledRejection` events,
when it should also track `rejectionHandled`? Sometimes it would cause
Jest to issue "thrown" warnings, and sometimes it would cause the entire
process to exit prematurely (before reporting results; especially
annoying when using watch mode).