Scriptable tool: wait for command to complete
The user may hit Done on the UI while the command is still running. We
want to be able to wait for the command to finish so that Scriptable
knows there is still stuff running.

We also try to abort any remaining fetches, but that doesn't help much
because the throttle has already pre-scheduled them to run on timers, so
it does not really speed things up much at all.

There is a Node/Scriptable inconsistency here in that if the fetcher
throws in Scriptable, it seems to cause the whole program to abort (a
limit's worth (10) pre-scheduled fetches still were started, but
`getISBNs`s Promise never resolved). This led me to making a Jest test
with a throwing fetcher and discovering the problem addressed in the
previous commit (`beforeAndAfter` implementation). I thought it might be
related to the Scriptable problem, but that seems to have turned out to
be throw vs. reject.