tool-core: abortable "editions of"
p-throttle has `.abort()`, that will reject the currently scheduled
calls, but we might still have stuff in the limit queue that feeds the
throttled fetchers.

p-limit has `.clearQueue()`, but it leaves any currently waiting
Promises unsettled (so our `Promise.allSettled` will never settle
either, leaving `getISBNs()` also unsettled).

So, generate a Promise internally and race it against the limited call's
Promise. We can clear the limiter's queue and reject this internal
Promise to let our Promises settle so that our functions finish. This
technically leaves unsettled (limiter-generated) Promises for unstarted,
limited calls, but there shouldn't be any more references to them (for
each unstarted call: one in the limiter's queue, and one in the Promise
race, but those would have been cleared or lost, respectively).

This also fixes the application of the throttles. The previous
implementation was unintentionally using separate throttles for each
Open Library service, when the intention was that they would share a
single throttle (since they are potentially hitting the same servers).