tool-core: refactored getISBNs and missingISBNs
Breaking up these functions will let interactive UIs do their own
filtering in the future (so, not just shelf-based filtering).

The testing is lacking, but since the core implementation is the same as
getISBNs and missingISBNs, it is covered by those tests.)If those
functions are ever removed, the their tests should be changed into tests
of the surviving functions.