csv: Row property value may be undefined
The previous definition implied that *every* string key produced a
string value, which was not true.

What I want is any string key that is present will have a string value.
Approximate that with any string key produces a value string value or
undefined.

This lets a Row have keys that are present, but have undefined as a
value, which is not really what I want, but is probably close enough.