Scriptable tool: revamp UI flow away from "commands"
The commands model came from reproducing the Node CLI tool. But, it
wasn't a great fit for an interactive UI like the Scriptable tool
offers.

So, instead of "missing ISBNs" and "get ISBNs", just read the data, let
the user select items (by shelf, for now), then show them the count of
items missing and not-missing ISBNs. They can save either of those, or
proceed with "editions" processing. "both ISBNs" is now an option when
saving either IBSN list.