Scriptable tool: help adjustments
Some details were lost in the earlier UI-flow revamp. Bring them back.

Also, reviewing and editing the array of strings (with manually broken
lines) was annoying. Switch to single strings with non-wrapped, long
lines and `outdent` to make it easier to review and edit (word wrapping
in-editor helps with the annoyingly long lines).