Scriptable tool: promisify UIRequestReceiver
Obviates the UI type, which was really just a collection of callbacks.

This should make it easier for the UI state redesign that I have in mind.

I was originally keeping the controller "at arms length" from the UI
with an eye towards making it easier to have them in separate execution
contexts (i.e. one in Scriptable and the other in a WebView). Such an
arrangement would involve having to make proxies on both sides, or at
least a controller proxy on the UI's side and callback proxies on the
controller's side. I already worked out a Scriptable->WebView
overlapping Promise proxy (evaluateJavaScript does not natively allow
overlapping asynchronously completed executions) that allows for
asynchronous "notifications" originating on the WebView side. It should
be possible to build atop that to make a Promise proxy in the reverse
direction, too. So, just give in and convert the controller to yielding
Promises and accept the future work of the WebView->Scriptable Promise
bridge if we ever want/need to put code in a WebView.