Scriptable tool: configuration screens
These screens are for adjusting state that may not be directly
accessible during the normal interaction (e.g. after something is
changed to a "do not warn/ask me again" state).

The network access permission are here (technically, they are
accessible, as long as the UI adds its row...), and the new
acknowledgement of "editions" being slow due to throttling.

Debug is move here, too, since it is a mostly out-of-the-way place.