papaparse: use Yarn patch instead of "re-bundling"
I am trying out "rollup-plugin-license", and having papaparse
"re-bundled" was causing it to be excluded from the gathered licenses. I
could have adopted the papaparse `package.json` "license" property and
LICENSE file, but then it would be wrong if they ever change the license
(probably unlikely, but important if it ever did happen).

Putting the bits that I need in a Yarn patch seems to work okay, and
should let the license info be directly visible. It will be brittle when
upgrading papaparse versions, but it might not be too bad to rebase the
patch every once in a while (until papaparse adopts an ESM solution that
Node understands is named-importable (like TS allows, given
@types/papaparse).