Scriptable tool: remove outdent options
The point was to keep potential CRLF line endings in `.js` files from
leaking into the resulting strings (since they would be used on iOS,
where LF is probably expected).

It turns out that this is a not a problem: JavaScript specifies that any
LF, CRLF, or CR instances in a template literal are always normalized to
LF. This holds for both plain and tagged template literals (and to both
the "cooked" and raw strings available in the tagged version).

TRV is used for the raw strings, its LineTerminatorSequence handling
specifies the normalization:

    https://tc39.es/ecma262/multipage/ecmascript-language-lexical-grammar.html#sec-static-semantics-trv

TV is used for "cooked" strings, it defers LineTerminatorSequence to TRV

    https://tc39.es/ecma262/multipage/ecmascript-language-lexical-grammar.html#sec-static-semantics-tv