Scriptable tool: change to allow full bundle-time outdenting
There was one instance where the plugin transformation declined to do
its bundle-time outdent transformation. It is due to a first-position
interpolated expression that the plugin does not analyze enough to know
it was not an indentation marker. Use an actual indentation marker to
make it explicit enough for the plugin to follow along.