rollup: bring preoutdent into top-level config
This took an ugly hack to let the Scriptable workspace's Rollup config
use the preoutdent workspace's generated module.

Just using `import()` is not enough, since the configurations are all
loaded before anything is built. We had to take it one step further and
defer loading the plugin until Rollup actually starts using it. The lets
Rollup build preoutdent before it is loaded when the Rollup starts
building its Scriptable work.