utils: provide export subpaths
Working on an installer that only needs the user agent helper from
`utils`.

Rollup was not able to tree-shake all the external packages out of that
scenario: even with moduleSideEffects:false it was keeping typanion
(probably due to the const definitions in the cache module).

So move to per-module exports to help Rollup ignore the irrelevant
stuff.

Also drop CommonJS generation. I am not using it and `utils` is pretty
unlikely to be used elsewhere (at least without breaking it up some).