Scriptable: automatic row height based on font and orientation
This is a squash-merge from the development branch (I don't want to take
the time to work it into publishable shape).

I wanted to automatically adjust the height of text-bearing UITable rows
to fit the text. Scriptable lets us specify the row height in points,
but there was no way to find out the point height for a given text cell.

The core idea here is to reproduce the built-in line breaking so that we
can compute the required height ourselves.

There a multiple steps to get there:

1. We need to know the width in points of area in which the text will
   eventually be drawn. This involves:

    * the screen width (`Device.screenSize().width`).
    * the undocumented "padding" applied to each UITableRow
    * the documented (but not in the context of Scriptable) safe area
      insets
    * the width used by other cells in the row (based on `widthWeight`)
    * the cell spacing

    It seems that Scriptable arranges for a row to have 40 points of
    width padding, plus whatever the safe area insets specify for the
    active presentation (usually zero, but non-zero when a notch/island
    is on the left or right). These figures were determined by fitting
    images into a UITableRow. They seem like they work okay, but it may
    not be entirely correct: text and image height is certainly handled
    differently in UITableRows.

2. Once we have the width that will be used for a text cell, we can
   attempt to reproduce the "native" line breaking (word wrapping) using
   measurements of the width of the font. Font measurement involves:

   * using DrawContext to create an image with some bit of text
   * shipping that image to a WebView
   * using the web Canvas API to look at individual pixels

   We "measure" various bits of text to measure the widths of certain
   characters, the average widths of other sets of characters, and the
   font's line spacing.

3. Now that we have a number of lines, we can just multiply it with the
   font's line spacing to find the text's height. Well, mostly. After
   measuring the row height required for N lines of text for every
   Dynamic Text size and every Scriptable "named" font, it turns out
   that we also need to add the "size" of the current body font (not the
   font used for the text!) before we will have enough row height for
   most fonts (a few require 1-4 more points, so we add 4 more points no
   matter which font or size we are using).

So that's enough to make a good guess at the required row height for a
given text cell. There are definitely edge cases that this will not
cover, but it seems to do fairly well for my English text use case.
Large fonts in narrow widths will be have line estimations that are way
too high (leaving lots of empty space in the row), and if a text use
lots of above-average width characters the estimate may end up being too
low (leaving the text truncated). There is room for improvement here:
measure the exact text instead of samples that get averaged, or measure
individual characters and cache them (probably font size changes are not
too common, but we don't have a good cache key for the Font used, since
Scriptable does not give us any details (name, size, etc.)).

But that isn't good enough. We need to adapt to orientation changes that
affect a row's available width. And we need to adapt to Dynamic Text
font changes (usually size, but also boldness) which affect character
widths. On and getting the safe area insets is a bit round-about, too:
the only place (that I know of) where they are accessible is through CSS
(another WebView to the rescue); the results are a bit flakey; I try to
patch them up.

Oh, and we use another layer of "UI runner" to watch for the
orientation, font, and inset changes and adjust the "builder" as
necessary.