Scriptable rollup: preserve Scriptable header, rotate color
When Scriptable finds a file that lacks a Scriptable header block, it
immediately adds one. So, when we write a new file via Rollup, then wait
for it up update in Scriptable on another device, iCloud Drive sees a
write from the PC, then a write from the device. We can avoid that
second write by preserving the Scriptable block.

We preserve everything except the color, which we rotate through a
more limited set to make it easier for me to notice changes.