Scriptable tool: do not offer config during progress
If the configuration state is activated from the progress state, the
next progress report (one every second or so) will switch back to the
progress state.

We could use a "sub" UI loop to "pause" the main UI loop, but it
probably isn't worth it (we would have to pass in the whole "UI" object,
not just the builder).

Don't offer the config button while in the progress state.