Scriptable check UITableRow: make measure-by-moire available
There are fewer steps to this process, but it is probably less reliable
because the very first step requires correctly choosing between patterns
that might be only subtlety different. If an incorrect choice is made it
will make the later steps meaningless or impossible.

I think the later steps are not as problematic (when the correct first
choice is made). They can actually serve as a kind of training in what
to look for in that first step. But, they are obviously useless for such
training since the first step has to be taken first. The core problem is
that the "first step" is dealing with two unknown paddings (height and
width) and thus can not even begin to control the scaling that happens
with an image size that is not exactly the size that will be displayed
unscaled.

Originally, the first "unknown" was determined in a different way, and
the second "unknown" could be reasonably determined by the moire
technique. But I discarded that original technique because it was hard
to design it to quickly settle on a correct value for any given device
size (it also required the use of the Zoom accessibility feature to
identify single-pixel width variations, which is maybe a bit much to
ask; the main "gaps" technique does rely on spotting possible
single-pixel gaps, but that turns out to be much easier than spotting
single-pixel width variations: "are there any high frequency changes in
the middle of this row" is easier to spot than "does the high frequency
change (at the end of this row) happen before or after that of the row
below").