Scriptable check UITableRow: show undefined vs. body font difference
The builder was previously using `Font.body()` if a titleFont was not
specified. We did this because body has (and the other named fonts have)
nice integer line spacings, while the "undefined font" has fractional
line spacings (looser at xSmall size, so prone to being truncated when
there are enough lines present), and different letter widths.

Add a "back door" way to actually leave the titleFont undefined so that
we can demonstrate the difference. At xSmall, the 50 lines will be
truncated on the "undefined" side because the height was calculated
based on measuring "body". At other sizes the "undefined" side will be
shorter than the "body" side, which isn't quite as problematic (not
causing truncation).