UITable-Runner: document drawText without setFont
I was hoping that drawing text without first setting a font would give
the equivalent font to an undefined titleFont. But the result does not
scale with Dynamic Type sizes and does not match any size of body font.
The measurement showed the same enWidth as xSmall body (and undefined),
but the lineSpacing was non-integral (like the undefined titleFont's
lineSpacing) but much tighter (16.7 vs. 19.1 (undefined titleFont) vs.
19 body).

Since this isn't a valid way to measure the undefined titleFont, we'll
keep defaulting to body in the builder.