apportionWidth: don't round weights
Demonstrated the zero widthWeight issue in the "check" app, so the
comment here can be deleted.

I tested some non-integer widthWeights, but didn't run into any
problems. The (unpublished, squashed) development history doesn't offer
any hints: the rounding was in place when it was first committed. So
remove the rounding until there is a documented reason to put it back.